Create an endpoint for route "/register"
This method will take name, email and password and store it in an
array as an object along with a random id.
To generate random id nodejs in-build module "crypto" is used.
For validation of user email and password "express validator" is used.
It will respond to user request with the registered user's profile data
that was stored in the array.